requirements.txt
anthropic>=0.40.0
pyairtable>=2.3.0
python-dotenv>=1.0.0
flask>=3.1.0
